+++ TankGWAR/TankGWAR/Texture.h
@@ -11,7 +11,7 @@ public:
 	~CTexture(void);
 	void Paint (RECT *,D3DXVECTOR2 *);
 	void Paint (D3DXVECTOR2 *);
-	void Paint (int,int);
+	void Paint (float,float);
 	void OnLostDevice();
 	int GetWidth();
 	int GetHeight();
